Move getAIMove function to goAI file
Function was being used by both the UI and the API, it did not belong in netscriptGoImplementation
